doc: rename markdown files to end with respective section
In a later change, we would like to support both `pandoc` and `ronn` as
`man-pages` generator for our project.

`ronn` will use source filename for man-pages `.TH` name and section.

Let's rename our Documentation source to prepare for that change.

While we're at it, unify release date in man pages to configure process.